// src/app/api/model/plato.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Plato { 
    id?: string;
    nombre?: string;
    description?: string;
    iconUrl?: string;
    tags?: Array<string>;
    possibleModifications?: Array<string>;
    precio?: number;
}